requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-generativeai
smtplib
email
